ART-52 Moved classes to a more adequate places
