+++ app/src/main/java/com/sefford/artdrian/Artpplication.kt
@@ -11,6 +11,7 @@ import com.sefford.artdrian.common.di.CoreModule
 import com.sefford.artdrian.common.di.DaggerApplicationComponent
 import com.sefford.artdrian.common.di.TopComponentHolder
 import com.sefford.artdrian.common.utils.debuggable
+import com.sefford.artdrian.connectivity.ConnectivityStore
 import com.sefford.artdrian.wallpapers.store.WallpaperEvents
 import com.sefford.artdrian.wallpapers.store.WallpaperStore
 import javax.inject.Inject
