+++ app/src/test/java/com/sefford/artdrian/wallpapers/domain/usecases/DownloadDtoWallpaperTestBuilder.kt
@@ -8,7 +8,7 @@ import kotlinx.coroutines.test.runTest
 import org.junit.jupiter.api.Test
 
 @OptIn(ExperimentalCoroutinesApi::class)
-class DownloadWallpaperTestBuilder {
+class DownloadDtoWallpaperTestBuilder {
 
     private lateinit var useCase: DownloadWallpaper
